test(assistant): shared-clone session integration locks
